replace embedded-time with fugit
`embedded-time` has some serious issues and moves too slowly and tries to do too much in single lib. Fugit is a time library that was built because some people were not happy to use embedded-time on embedded systems.
Fugit only deals with time conversions and those conversions are done using const generics. So most of the stuff is  happening during compilation.
This is very light weight compared to embedded-time.