Replace embedded-time with fugit #54 (#57)
* Cargo.toml: replace embedded-time with fugit

And use ublox-sockets that support fugit time crate.

* lib: introduce new Clock trait and Instant alias

* replace embedded-time with fugit

`embedded-time` has some serious issues and moves too slowly and tries to do too much in single lib. Fugit is a time library that was built because some people were not happy to use embedded-time on embedded systems.
Fugit only deals with time conversions and those conversions are done using const generics. So most of the stuff is  happening during compilation.
This is very light weight compared to embedded-time.

* rename FREQ_HZ to TIMER_HZ

the same is used in rtic and it is more clear what it is

* lib: export Clock from atat

* use Clock trait from atat lib

* examples: update common_lib to latest

* examples: linux wip

* Make sure the embedded-nal implementations only makes use of error-ty… (#53)

* Make sure the embedded-nal implementations only makes use of error-types from ublox-sockets

* Add embedded-hal-driver keyword to crates.io
# Conflicts:
#	ublox-cellular/Cargo.toml
#	ublox-cellular/src/services/data/tcp_stack.rs
#	ublox-cellular/src/services/data/udp_stack.rs

* test_helpers: update to WouldBlock based Clock

* error: add helper to convert Clock error to Error::Generic

From trait cannot be used here because Clock error is associated type

* client: use nb::block to wait for timer

* import atat from blackbirdhq master

* use fugit defmt feature to get Format impl

* linux: set_socket_storage

* add IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr to prelude

* linux example: add socket read/write

* reexport fugit

fugit is needed for implementing Clock trait therefore it makes sense to reexport it here

* add some general documentation about the crate

* make first examples buildable also on macos

on macos dummy gpio support is used

* rename linux example to sockets, add args support

* power: add nb::block! for blocking timer.waits

without those it returns immediately without blocking in place

* use atat and ublox-sockets from crates.io

* use fmt.rs to support log or defmt

* examples: add better logging

* update to embedded-hal 1.0.0-alpha.6

* add .DS_Store to .gitignore

macOS creates those nasty files

* update atat and ublox-sockets

* Update ublox-cellular/Cargo.toml

* Replace CharVec with Bytes, following deprecation warning in ATAT

* Fix tests & add defmt-impl feature

Co-authored-by: Andres Vahter <andres@vahter.me>
Co-authored-by: Andres Vahter <andres.vahter@gmail.com>